refactor: move test-only helpers to tests
